+++ actions/UpdateAction.php
@@ -68,6 +68,11 @@ class UpdateAction extends Action
 
         $model = $this->model;
 
+        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post()) && isset($_POST['ajax'])) {
+            Yii::$app->response->format = Response::FORMAT_JSON;
+            return \yii\widgets\ActiveForm::validate($model);
+        }
+
         if ($model->load(Yii::$app->request->post())) {
             if ($model->save()) {
 
@@ -92,40 +97,6 @@ class UpdateAction extends Action
                 'message' => 'error',
                 'error' => print_r($model->getErrors()),
             ];
-        }
-
-        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post()) && isset($_POST['ajax'])) {
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return \yii\widgets\ActiveForm::validate($model);
-        }
-
-        return $this->controller->renderAjax($this->view, [
-            'model' => $model
-        ]);
-
-        $model = $this->model;
-
-        if ($model->load(Yii::$app->request->post())) {
-            if ($model->save()) {
-
-                Yii::$app->getSession()->setFlash('success', $this->successMsg);
-                Yii::$app->response->format = Response::FORMAT_JSON;
-                return [
-                    'message' => 'success',
-                ];
-            }
-
-            Yii::$app->getSession()->setFlash('error', $this->errorMsg);
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return [
-                'message' => 'error',
-                'error' => print_r($model->getErrors()),
-            ];
-        }
-
-        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post()) && isset($_POST['ajax'])) {
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return ActiveForm::validate($model);
         }
 
         return $this->controller->renderAjax($this->view, [